separate ScrollDettector from TimetableView
